Refactor Code
Reorganized components
